styling the game and some refactoring
