+++ package.json
@@ -3,11 +3,13 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@material-ui/core": "^4.11.0",
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.3.2",
     "@testing-library/user-event": "^7.1.2",
     "axios": "^0.20.0",
     "colorette": "^1.2.1",
+    "fontsource-roboto": "^3.0.3",
     "lodash": "^4.17.20",
     "react": "^16.13.1",
     "react-dice-complete": "^1.2.1",
