--- src/App.js
@@ -1,13 +1,0 @@
-import React from 'react';
-import GamesCreator from "./GamesCreator";
-
-function App() {
-  return (
-      <div>
-        <GamesCreator/>
-      </div>
-  );
-}
-
-
-export default App;
